refactor: AutoResizeTextarea 대신 InpuTextarea 공통 컴포넌트 사용
