+++ src/components/common/Input/InputTextarea/index.tsx
@@ -8,12 +8,14 @@ interface InputTextareaProps
   variant?: 'base' | 'box' | 'reply'; // 할 일 댓글 수정 | 자유게시판 | 할 일 댓글 작성
   className?: string;
   onClick?: () => void;
+  isSubmitDisabled: boolean;
 }
 
 const InputTextarea = ({
   variant,
   className = '',
   onClick,
+  isSubmitDisabled = false,
   ...props
 }: InputTextareaProps) => {
   const textareaRef = useRef<HTMLTextAreaElement>(null);
@@ -38,18 +40,19 @@ const InputTextarea = ({
           variant === 'box' &&
             'rounded-[12px] border border-slate-50/10 bg-slate-800',
 
-          variant === 'reply' && 'border-t border-b border-slate-50/10 py-3',
+          variant === 'reply' && 'border-t border-b border-slate-50/10 py-3.5',
           className
         )}
       />
       {variant === 'reply' && (
         <button
           onClick={onClick}
+          disabled={isSubmitDisabled}
           className={clsx(
             'absolute top-3 right-0',
             'flex items-center justify-center',
             'size-6 rounded-full bg-green-700',
-            'transition-colors duration-100 hover:bg-green-800 focus:bg-green-900 active:bg-green-900',
+            'transition-colors duration-100 hover:bg-green-800',
             'disabled:cursor-default disabled:bg-slate-400'
           )}
         >
